requirements.txt
aiogram~=3.5.0
pydantic~=2.7.1
pydantic_settings
django-scheduler~=0.10.1